requirements.txt
asgiref==3.7.2
black==24.2.0
click==8.1.7
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
